refactor: jobField1라고 쓰인 불용어 제거
