+++ src/main/java/kakaotech/bootcamp/respec/specranking/domain/spec/spec/dto/cache/CachedRankingDto.java
